% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_merge_data_bitcoincharts.R
\name{aggregate_merge_bictoincharts_data}
\alias{aggregate_merge_bictoincharts_data}
\title{Aggregate and merge time series data downloaded from bitcoincharts}
\usage{
aggregate_merge_bictoincharts_data(
  data_list,
  level_of_aggregation = "minutes",
  aggregate_every = 15
)
}
\arguments{
\item{data_list}{is a list containing the dataframes of several time series data downloaded from bitcoincharts}

\item{level_of_aggregation}{can be 'seconds', 'minutes' or 'hours'}

\item{aggregate_every}{defines how many intervals to aggregate (depends on level_of_aggregation)}
}
\value{
results a list containing the merged intraday bitcoin prices (an xts object) and their daily returns (an xts object)
}
\description{
This function aggregates and merges time series data downloaded from bitcoincharts
}
\details{
This function aggregates and merges time series data downloaded from bitcoincharts and provided to the function as a list.
It returns a list containing the dataframe with the merged price data according to the desired level of aggregation at the intraday level,
as well as a dataframe with the merged daily data returns.
}
\examples{
\dontrun{
library(highfrequency)
localbtcMXN <-bitcoincharts_single_download(name = "localbtcMXN.csv.gz")
btcdeEUR <-bitcoincharts_single_download(name = "btcdeEUR.csv.gz")
krakenEUR <-bitcoincharts_single_download(name = "krakenEUR.csv.gz")
localbtcINR <-bitcoincharts_single_download(name = "localbtcINR.csv.gz")
coinbaseUSD <-bitcoincharts_single_download(name = "coinbaseUSD.csv.gz")
data5<-list(localbtcMXN =localbtcMXN, btcdeEUR=btcdeEUR, krakenEUR=krakenEUR,
             localbtcINR=localbtcINR, coinbaseUSD=coinbaseUSD)
data_clean<-aggregate_merge_bictoincharts_data(data_list=data5)
}

}
