% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitcoincharts_download_large.R
\name{bitcoincharts_download_large}
\alias{bitcoincharts_download_large}
\title{Download a pre-specified bitcoin currency pair time series from api.bitcoincharts.com}
\usage{
bitcoincharts_download_large(
  address = "D:/Downloads/",
  name = "bitchangePLN.csv.gz",
  save_data = FALSE
)
}
\arguments{
\item{address}{is the directory address where to save the data}

\item{name}{is the name of the file to be downloaded from api.bitcoincharts.com}

\item{save_data}{if TRUE, then the downloaded trades price series will be saved in the bitcoin_data.rds file (in the Working Directory)}
}
\value{
data is a dataframe containing 'timestamp', 'price', 'volume' for the selected bitcoin trades time series.
}
\description{
This function downloads a pre-specified bitcoin currency pair time series from api.bitcoincharts.com and it is suitable for large files
}
\details{
This function downloads a pre-specified bitcoin currency pair time series  from api.bitcoincharts.com .
It is better to use this function for downloading large files instead of bitcoincharts_single_download,
which may gives an error (depending on the platform used).
}
\examples{
\dontrun{
bb <-bitcoincharts_download_large(name = "coinbaseEUR.csv.gz", save_data=TRUE)
str(bb)
}

}
