% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data_bitcoincharts.R
\name{bitcoincharts_random_download}
\alias{bitcoincharts_random_download}
\title{Download a pre-specified number of random bitcoin trades time series from api.bitcoincharts.com}
\usage{
bitcoincharts_random_download(number_of_series_needed = 5,
  exceptions = c("okcoinCNY.csv.gz", "coincheckJPY.csv.gz"),
  start_date = "2015-08-03", end_date = "2017-07-01",
  save_data = FALSE)
}
\arguments{
\item{number_of_series_needed}{is the desidered number of (randomly selected) bitcoin trades series}

\item{exceptions}{is a string vector of the files not to be (potentially) selected from api.bitcoincharts.com due to their large dimensions}

\item{start_date}{this is the minimum starting date required}

\item{end_date}{this is the maximum end date required}

\item{save_data}{if TRUE, then the downloaded trades series will be saved in the bitcoin_data.rds file (in the Working Directory)}
}
\value{
data_list is a list containting the dataframes of several time series data downloaded from bitcoincharts. Each dataframe has
the 'timestamp', 'price', 'volume' for the selected bitcoin trades time series.
}
\description{
This function downloads a pre-specified number of random bitcoin trades series  from api.bitcoincharts.com
}
\details{
This function downloads a pre-specified number of random bitcoin trades time series from api.bitcoincharts.com with a
required minimum starting date and a required ending date
}
\examples{
\dontrun{
aa <-bitcoincharts_random_download(number_of_series_needed = 3, save_data=TRUE)
str(aa)
}

}
