% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data_bitcoincharts.R
\name{bitcoincharts_single_download}
\alias{bitcoincharts_single_download}
\title{Download a pre-specified bitcoin currency pair time series from api.bitcoincharts.com}
\usage{
bitcoincharts_single_download(name = "bitstampUSD.csv.gz",
  start_date = "2015-08-03", end_date = "2017-07-01",
  save_data = FALSE)
}
\arguments{
\item{name}{is the name of the file to be downloaded from api.bitcoincharts.com}

\item{start_date}{this is the minimum starting date required}

\item{end_date}{this is the maximum end date required}

\item{save_data}{if TRUE, then the downloaded trades price series will be saved in the bitcoin_data.rds file (in the Working Directory)}
}
\value{
data is a dataframe containing 'timestamp', 'price', 'volume' for the selected bitcoin trades time series.
}
\description{
This function downloads a pre-specified bitcoin currency pair time series from api.bitcoincharts.com
}
\details{
This function downloads a pre-specified bitcoin currency pair time series  from api.bitcoincharts.com with a
required minimum starting date and a required ending date
}
\examples{
\dontrun{
bb <-bitcoincharts_single_download(name = "coinbaseEUR.csv.gz", save_data=TRUE)
str(bb)
}

}
