% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_bitcoinity.R
\name{bitcoinity_download}
\alias{bitcoinity_download}
\title{Download the time series data of an exchange pair involving bitcoin and a fiat currency.}
\usage{
bitcoinity_download(
  currency = "USD",
  data_type = "price_volume",
  exchange = "bitstamp",
  freq = "hour",
  time_length = "30d",
  sd = NULL
)
}
\arguments{
\item{currency}{is the quote currency of the currency pair}

\item{data_type}{is the type of data required. Possible cases are: "volume","rank","price",
"price_volume","market_cap","tradespm","volatility","spread","bidask_sum".}

\item{exchange}{is the cryptocurrency exchange}

\item{freq}{is the data frequency: "hour", "day", or "week"}

\item{time_length}{is the time length required: "10m","1h","6h","24h", "3d", "7d", "30d", "6m", "2y", "5y", "all".}

\item{sd}{is the spread depth. Possible values in BTC are: NULL (default), "10", "100".}
}
\value{
a dataframe with the requested data
}
\description{
This function downloads the time series data of an exchange pair involving bitcoin and a fiat currency,
given the specified quote currency, data type, cryptocurrency exchange, data frequency, time length,
spread depth (if requested).
}
\examples{
dat<-bitcoinity_download(currency="USD", data_type="spread", exchange=NULL,sd="10")
head(dat)[1:4]
}
