% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_bitcoinity.R
\name{bitcoinity_download_DIK}
\alias{bitcoinity_download_DIK}
\title{Download the time series data of the depth for price impact, for an exchange pair involving bitcoin and a fiat currency.}
\usage{
bitcoinity_download_DIK(
  bp = "10",
  bu = "b",
  currency = "USD",
  data_type = "bidask_sum",
  exchange = "bitstamp",
  freq = "hour",
  time_length = "30d"
)
}
\arguments{
\item{bp}{is the price range (k) for the depth for price impact}

\item{bu}{can be "b" (BTC) or "c" (currency). If you choose the former, the depth sum is expressed in BTC;
if you choose the latter, the depth sum is expressed in the quote currency.}

\item{currency}{is the quote currency of the currency pair}

\item{data_type}{is the type of data required. Possible cases are: "volume","rank","price",
"price_volume","market_cap","tradespm","volatility","spread","bidask_sum".}

\item{exchange}{is the cryptocurrency exchange}

\item{freq}{is the data frequency: "hour", "day", or "week"}

\item{time_length}{is the time length required: "10m","1h","6h","24h", "3d", "7d", "30d", "6m", "2y", "5y", "all".}
}
\value{
a dataframe with the requested data
}
\description{
This function downloads the time series data of the depth for price impact, for an exchange pair involving bitcoin and a fiat currency,
given the specified price range in percentage (k), depth sum expressed in BTC or in quote currency, quote currency, data type,
cryptocurrency exchange, data frequency, time length.
}
\examples{
aa <- bitcoinity_download_DIK(bp="10",bu="b",currency="USD",data_type="bidask_sum",
      exchange="bitstamp",freq="hour",time_length="30d")
head(aa)
}
