% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_bound.R
\name{btc.arbitrage.user}
\alias{btc.arbitrage.user}
\title{Compute the theoretical no-arbitrage market price of an altcoin based on the SHA-256 algorithm}
\usage{
btc.arbitrage.user(
  block.reward.bitcoin = 12.5,
  hashing.power.miner = 10^12,
  Difficulty.bitcoin = 922724699726,
  block.reward.altcoin = 12.5,
  Difficulty.altcoin = 109634813602
)
}
\arguments{
\item{block.reward.bitcoin}{is the block reward for bitcoin (currently 12.5 BTC/block)}

\item{hashing.power.miner}{is the hashing power employed by a miner. Hayes (2017) sets this value to 1000 GH/s even though
the actual hashing power of a miner is likely to deviate greatly from this value.}

\item{Difficulty.bitcoin}{is the difficulty of the Bitcoin system which is expressed in units of GH/block}

\item{block.reward.altcoin}{is the block reward for the SHA-256 based altcoin}

\item{Difficulty.altcoin}{is the difficulty of the altcoin system which is expressed in units of GH/block}
}
\value{
epsilon_star is no-arbitrage market price of an altcoin based on the SHA-256 algorithm
}
\description{
This function computes the theoretical no-arbitrage market price of an altcoin based on the SHA-256 algorithm
}
\details{
This function computes the no-arbitrage market price of an altcoin based on the SHA-256 algorithm
}
\examples{
#no-arbitrage market price of Bitcoin Cash using data published on the 16/09/2017
 btc.arbitrage.user(block.reward.bitcoin = 12.5, hashing.power.miner = 10^12,
 Difficulty.bitcoin = 922724699726,
 block.reward.altcoin = 12.5,  Difficulty.altcoin = 109634813602)
}
\references{
Hayes, A. S. (2017). Cryptocurrency Value Formation: An empirical study leading to a cost of production model for valuing Bitcoin. Telematics and Informatics, forthcoming.

Fantazzini, D., Nigmatullin, E., Sukhanovskaya, V., Ivliev, S., (2016) Everything you always wanted to know about bitcoin modelling but were afraid to ask. Part 1. Applied Econometrics, 44, 5-24.
}
