% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_bound.R
\name{btc.lower.bound.user}
\alias{btc.lower.bound.user}
\title{Compute the bitcoin lower bound with user-delivered inputs}
\usage{
btc.lower.bound.user(
  block.reward = 12.5,
  hashing.power.miner = 10^12,
  Difficulty = 559970892890,
  price.kWh = 0.125,
  W.GHs = 0.25
)
}
\arguments{
\item{block.reward}{is the block reward (currently 12.5 BTC/block)}

\item{hashing.power.miner}{is the hashing power employed by a miner. Hayes (2017) sets this value to 1000 GH/s even though
the actual hashing power of a miner is likely to deviate greatly from this value.}

\item{Difficulty}{is the difficulty which is expressed in units of GH/block}

\item{price.kWh}{is the $electricity price per kWh in dollars}

\item{W.GHs}{is the energy consumption efficiency of the producer’s hardware expressed in W per GH/s}
}
\value{
price.bitcoin is the bitcoin lower bound
}
\description{
This function computes the bitcoin lower bound with inputs given by the user and not taken from the web
}
\details{
This function computes the bitcoin lower bound with inputs given by the user
}
\examples{
#This reproduces the example in Fantazzini et al. (2016)
 btc.lower.bound.user(Difficulty=60883825480, price.kWh=0.135,W.GHs=0.75)
 #  588.3616

}
\references{
Hayes, A. S. (2017). Cryptocurrency Value Formation: An empirical study leading to a cost of production model for valuing Bitcoin. Telematics and Informatics, forthcoming.

Fantazzini, D., Nigmatullin, E., Sukhanovskaya, V., Ivliev, S., (2016) Everything you always wanted to know about bitcoin modelling but were afraid to ask. Part 1. Applied Econometrics, 44, 5-24.
}
