% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upper_bounds.R
\name{btc.upper.bound.web}
\alias{btc.upper.bound.web}
\title{Compute the bitcoin upper bound with web-delivered inputs}
\usage{
btc.upper.bound.web(fredkey, start.date = "2004-01-01",
  end.date = "2014-12-31", lastyear = NULL)
}
\arguments{
\item{fredkey}{is the FRED API KEY (requires registration)}

\item{start.date}{is the starting date (format "%Y-%m-%d") for downloading the US household checking deposits and cash and
US personal consumption data from the FRED database}

\item{end.date}{is the ending date (format "%Y-%m-%d") for downloading the US household checking deposits and cash and
US personal consumption data from the FRED database}

\item{lastyear}{represents the last year to be used for computing the upper bound. If NULL, the last available data for each
categories will be used (which may be different across categories).Otherwise a common year equal to lastyear
will be used for all categories: in this case be sure the the common year is available for all the 3 categories used
to compute the upper bound, otherwise you will get an error.}
}
\value{
btc.upper.bound.web is the bitcoin upper bound
}
\description{
This function computes the bitcoin upper bound with inputs taken from the web
}
\details{
This function computes the bitcoin upper bound with inputs taken from the web
}
\examples{
\dontrun{
 #Here insert your valid FRED key
 fredkey ="5565e828f27269b6ed4f3c2552d0cc8f"
 btc.upper.bound.web(fredkey)
}
}
