% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liquidity_cryptowat.R
\name{calcBestOrdersAndTradesBased}
\alias{calcBestOrdersAndTradesBased}
\title{Calculate liquidity measures  based on best orders and last trade}
\usage{
calcBestOrdersAndTradesBased(market, pair, sourceDirectory, destDirectory)
}
\arguments{
\item{market}{is the required exchange}

\item{pair}{is the required exchange rate pair}

\item{sourceDirectory}{is the directory path where the matched.csv file is located}

\item{destDirectory}{is the directory path where the matched_liquidity_measures.csv file will be saved}
}
\value{
df is a dataframe with liquidity measures based on best orders and last trades
}
\description{
This function calculates liquidity measures based on best orders and last trade
}
\details{
This function calculates liquidity measures based on best orders and last trade
}
\examples{
\dontrun{
data_file<-system.file("extdata", package = "bitcoinFinance")
prepareTradesFromDirectory("coinbase","btcusd",data_file,data_file)
prepareOrders("coinbase","btcusd",data_file,data_file)
prepareBestOrders ("coinbase","btcusd",data_file )
lastPriceToBestOrders ("coinbase","btcusd",data_file )
trades_liquidity<-calcTradeBased("coinbase","btcusd",data_file, data_file )
bestorders_liquidity<-calcBestOrderBased("coinbase","btcusd",data_file, data_file )
bestorders_liquidity_trades<-calcBestOrdersAndTradesBased("coinbase","btcusd",data_file, data_file )
}
}
