% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liquidity_cryptowat.R
\name{calcTradeBased}
\alias{calcTradeBased}
\title{Calculate liquidity measures based on trades}
\usage{
calcTradeBased(market, pair, sourceDirectory, destDirectory)
}
\arguments{
\item{market}{is the required exchange}

\item{pair}{is the required exchange rate pair}

\item{sourceDirectory}{is the directory path where the  trades.csv file is located}

\item{destDirectory}{is the directory path where the trades_liquidity_measures.csv file will be saved}
}
\value{
newData is a dataframe with the trade-based liquidity measures
}
\description{
This function calculates liquidity measures based on trades
}
\details{
This function calculates liquidity measures based on trades
}
\examples{
\dontrun{
data_file<-system.file("extdata", package = "bitcoinFinance")
prepareTradesFromDirectory("coinbase","btcusd",data_file,data_file)
prepareOrders("coinbase","btcusd",data_file,data_file)
prepareBestOrders ("coinbase","btcusd",data_file )
lastPriceToBestOrders ("coinbase","btcusd",data_file )
trades_liquidity<-calcTradeBased("coinbase","btcusd",data_file, data_file )
}
}
