% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_coindesk_bpi.R
\name{coindesk_download}
\alias{coindesk_download}
\title{Download the daily time series data of the CoinDesk Bitcoin Price Index (BPI).}
\usage{
coindesk_download(start = "2011-01-01", end = "2017-07-27")
}
\arguments{
\item{start}{is the starting point of the required daily data sample}

\item{end}{is the ending point of the required daily  data sample}
}
\value{
a xts object containing the required historical daily data for the CoinDesk Bitcoin Price Index
}
\description{
This function downloads the daily time series data of the CoinDesk Bitcoin Price Index (BPI),
given a specified starting point and ending point.
}
\examples{
\dontrun{
xbp_data <- coindesk_download(start="2011-01-01",end="2017-07-27")
}
}
