% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liquidity_cryptowat.R
\name{cryptowatch_data_download}
\alias{cryptowatch_data_download}
\title{Download trades data or order book data from cryptowat.ch}
\usage{
cryptowatch_data_download(
  exchange,
  pair,
  query_type,
  query_options = NULL,
  data_dir = NULL,
  file_prefix = NULL,
  logfile = "loader.log",
  register = "register.csv"
)
}
\arguments{
\item{exchange}{is the required exchange}

\item{pair}{is the required exchange rate pair}

\item{query_type}{can be either "orderbook" or "trades"}

\item{query_options}{can be either NULL or "limit=100" (in the latter case the limit can vary)}

\item{data_dir}{is the file path where to save the downloaded data}

\item{file_prefix}{is a prefix to the filename where the data will be saved.If NULL, date and time will be added.}

\item{logfile}{is a text file where the situation of the download process will be saved}

\item{register}{collects the names of the successfully downloaded files and their characteristics}
}
\description{
This function downloads trades data or order book data from cryptowat.ch
}
\details{
This function downloads trades data or order book data from cryptowat.ch and save them in a .json file
}
\examples{
\dontrun{
cryptowatch_data_download("coinbase","btcusd", "orderbook", data_dir=getwd() )
cryptowatch_data_download("coinbase","btcusd", "trades",
   query_options="limit=100", data_dir=getwd() )
}
}
