% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_coinmarketcap.R
\name{download_coinmarketcap_daily}
\alias{download_coinmarketcap_daily}
\title{Download the daily data of the selected cryptocurrency from coinmarketcap.com}
\usage{
download_coinmarketcap_daily(cryptoname = "bitcoin",
  start_date = 20130428, end_date = 20170705)
}
\arguments{
\item{cryptoname}{is the selected cryptocurrency}

\item{start_date}{is the starting date (format YYYYMMDD) for downloading the selected cryptocurrency daily data}

\item{end_date}{is the ending date (format YYYYMMDD) for downloading the selected cryptocurrency daily data}
}
\value{
table_news is a dataframe with daily open high low close turnover and market capitalization
}
\description{
This function downloads the daily data of the selected cryptocurrency from coinmarketcap.com
}
\details{
This function downloads the daily data of the selected cryptocurrencies from coinmarketcap.com
}
\examples{
download_coinmarketcap_daily(cryptoname="ethereum", end_date=20170704)

}
