% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_coinmarketcap.R
\name{download_coinmarketcap_w}
\alias{download_coinmarketcap_w}
\title{Download the weekly data of the selected cryptocurrency from coinmarketcap.com (ARCHIVE ONLY - DO NOT USE IT)}
\usage{
download_coinmarketcap_w(
  id = 1,
  currency = "USD",
  start_date = "2018-09-16",
  end_date = "2020-09-23"
)
}
\arguments{
\item{id}{is the coinmarketcap ID used for the selected cryptocurrency (see the coinmarketcap API for the full ID list)}

\item{currency}{is the reference currency used in the crypto pair}

\item{start_date}{is the starting date (format YYYY-MM-DD) for downloading the selected cryptocurrency weekly data}

\item{end_date}{is the ending date (format YYYY-MM-DD) for downloading the selected cryptocurrency weekly data}
}
\value{
table_news is a dataframe with weekly open high low close volume and market capitalization
}
\description{
This function downloaded (before 2024) of the selected cryptocurrency from coinmarketcap.com
}
\details{
ATTENTION: THE COINMARKETCAP API DOES NOT WORK ANY MORE SINCE EARLY 2024. DO NOT USE THESE FUNCTIONS.
THESE FUNCTIONS ARE LEFT IN THE PACKAGE ONLY FOR ARCHIVE AND EDUCATIONAL PURPOSES ONLY.
THE QUICKEST WAY TO DOWNLOAD FREE CRYPTO DATA IS BY USING www.cryptodatadownload.com .
This function downloaded (before 2024)  the weekly data of the selected cryptocurrencies from coinmarketcap.com
}
