% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_coinmarketcap.R
\name{download_coinmarketcap_weekly}
\alias{download_coinmarketcap_weekly}
\title{Download the weekly data of the selected cryptocurrency from coinmarketcap.com}
\usage{
download_coinmarketcap_weekly(dates, cryptoname)
}
\arguments{
\item{dates}{is a sequence of dates (format YYYYMMDD) for downloading the selected cryptocurrency weekly data}

\item{cryptoname}{is the selected cryptocurrency}
}
\value{
price is a numeric vector of weekly prices
}
\description{
This function downloads the weekly data of the selected cryptocurrency from coinmarketcap.com
}
\details{
This function downloads the weekly data of the selected cryptocurrencies from coinmarketcap.com .
Warning: this function is quite time-consuming becasue it opens a weekly webpage one after the other.
It is used here for educational purposes only.
}
\examples{
dates<-seq(as.Date("2015/08/09"), by = "week", length.out = 3)
dates <- as.numeric( gsub("-", "", dates) )
cryptoname <-c("bitcoin", "ethereum")
coins_all=NULL
for (nn in cryptoname) {
  print(nn)
  pcoin<-download_coinmarketcap_weekly(dates, nn)
  coins_all=cbind(coins_all,pcoin)
}
coins_all=cbind(dates, coins_all)
colnames(coins_all)=c("dates",cryptoname)
as.data.frame(coins_all)

}
