% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liquidity_cryptowat.R
\name{get_time_str}
\alias{get_time_str}
\title{This is a small function to get a time string}
\usage{
get_time_str(separator = "_")
}
\arguments{
\item{separator}{a character string to separate the terms}
}
\value{
a character time string for the current date and time
}
\description{
This is a small function to get a time string
}
\details{
This function delivers a character time string for the current date and time
}
\examples{
get_time_str()
}
