% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liquidity_cryptowat.R
\name{getFilenames}
\alias{getFilenames}
\title{Get the file names of data located in a specified directory and with certain requirements}
\usage{
getFilenames(market, pair, type, file_extension, sourceDirectory)
}
\arguments{
\item{market}{is the required exchange}

\item{pair}{is the required exchange rate pair}

\item{type}{can be either "orderbook" or "trades"}

\item{file_extension}{is the file extension of the required dataset}

\item{sourceDirectory}{is the directory path where the data files are located}
}
\value{
needed_files a list of all files
}
\description{
This function gets the file names of data located in a specified directory and with certain requirements
}
\details{
This function gets the file names of data located in a specified directory and containing
certain keywords (for market, currency pair, data type), and file extension
}
\examples{
data_file<-system.file("extdata", package = "bitcoinFinance")
filenames_trades <- getFilenames("coinbase","btcusd","trades",".json", data_file )
filenames_orderbook <- getFilenames("coinbase","btcusd","orderbook",".json", data_file )
}
