% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data_poloniex.R
\name{\%p\%}
\alias{\%p\%}
\title{Pipe function to unite two characters}
\usage{
e1 \%p\% e2
}
\arguments{
\item{e1}{is the first term to be united}

\item{e2}{is the second term to be united}
}
\description{
Pipe function to unite two characters
}
\examples{
"United " \%p\% "States"
}
