% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HARRV_recursive_forecast.R
\name{HARRV.recursive.forecast}
\alias{HARRV.recursive.forecast}
\title{Compute the recursive forecasts for the HAR-RV model.}
\usage{
HARRV.recursive.forecast(
  dat,
  periods = c(1, 5, 22),
  type = "HARRV",
  transform = NULL,
  roll.window = 1621,
  h = 1
)
}
\arguments{
\item{dat}{is a xts object containing intraday 5-minute regularly spaces prices (see example below)}

\item{periods}{is a vector of integers indicating over how days the realized measures in the model should be aggregated.
By default periods = c(1,5,22). It is needed for the computation of the \link[highfrequency]{HARmodel} of the \code{highfrequency} package}

\item{type}{is a string referring to the type of HAR model you would like to estimate using \link[highfrequency]{HARmodel}. By default type = "HARRV".}

\item{transform}{optionally a string referring to a function that transforms both the dependent and explanatory variables in the \link[highfrequency]{HARmodel}.
By default transform=NULL, so no transformation is done. Typical other choices in this context would be "log" or "sqrt".}

\item{roll.window}{is the rolling window size used for estimating the HAR-RV model}

\item{h}{is the desired forecasting horizon. Default is 1 (that is, 1-step-ahead recursive forecasts)}
}
\value{
roll.fore a xts object containing the h-day-ahead RV forecasts and the realized RV forecasts.
}
\description{
This function computes the recursive forecasts for the HAR-RV model
}
\details{
This function computes the recursive forecasts for the HAR-RV model.
See Chevillon (2007, 2016) for a discussion about direct and recursive forecasts.
}
\examples{
\dontrun{
library(highfrequency)
aa <-bitcoincharts_single_download(name = "bitstampUSD.csv.gz", save_data=TRUE)
data_clean<-aggregate_merge_bictoincharts_data(data_list=list(aa=aa), aggregate_every = 5)
dat<-data_clean$price_ts
dat<- dat["2013-01-02/2017-07-12"]
bb<-HARRV.recursive.forecast(dat=dat)
tail(bb)
 #                     predicted     realized
 # 2017-07-07 23:55:00 0.0008078943 0.003320551
 # 2017-07-08 23:55:00 0.0010431969 0.002369109
 # 2017-07-09 23:55:00 0.0008631436 0.002488993
 # 2017-07-10 23:55:00 0.0021875888 0.002373416
 # 2017-07-11 23:55:00 0.0039407938 0.003152035
 # 2017-07-12 23:55:00 0.0025053951 0.004210418
}
}
\references{
Chevillon, G. (2007). Direct multi-step estimation and forecasting. Journal of Economic Surveys,21(4),746-785.

Chevillon, G. (2016). Multistep forecasting in the presence of location shifts. International Journal of Forecasting, 32(1), 121-137.
}
