% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liquidity_cryptowat.R
\name{lastPriceToBestOrders}
\alias{lastPriceToBestOrders}
\title{Match last trades data with best bids and best asks}
\usage{
lastPriceToBestOrders(market, pair, sourceDirectory)
}
\arguments{
\item{market}{is the required exchange}

\item{pair}{is the required exchange rate pair}

\item{sourceDirectory}{is the directory path best_orders.csv and trades.csv files are located
(and where the matched.csv file will be saved)}
}
\description{
This function matches last trades data with best bids and best asks
}
\details{
This function matches last trades data (from trades.csv) with best bids and
 best asks (from best_orders.csv) and save the matched data in the matched.csv file
}
\examples{
\dontrun{
data_file<-system.file("extdata", package = "bitcoinFinance")
prepareTradesFromDirectory("coinbase","btcusd",data_file,data_file)
prepareOrders("coinbase","btcusd",data_file,data_file)
prepareBestOrders ("coinbase","btcusd",data_file )
lastPriceToBestOrders ("coinbase","btcusd",data_file )
}
}
