% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_pd.R
\name{matrix_pd}
\alias{matrix_pd}
\title{Check whether a matrix is positive definite and regularize it if not positive definite}
\usage{
matrix_pd(sym_matrix)
}
\arguments{
\item{sym_matrix}{is a symmetric matrix}
}
\value{
cxpd a (potentially regularized) positive definite a symmetric \code{n} x \code{n} matrix
}
\description{
This function checks whether a matrix is positive definite and regularizes it if not positive definite
}
\details{
This function checks whether a matrix is positive definite and regularizes it if not positive definite,
as done in  Hautsch, N., L. Kyj, and R. Oomen (2012)
}
\examples{
a = rbind(
   c(4,  0.5,1),
   c(0.5, 3 ,0),
   c(1,   0, -2)
   )
a
matrix_pd(a)
}
\references{
Hautsch, N., L. Kyj, and R. Oomen (2012). A blocking and regularization approach
to high-dimensional realized covariance estimation. Journal of Applied Econometrics, 27, 625–645
}
