% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data_poloniex.R
\name{poloniex_download}
\alias{poloniex_download}
\title{Download time series data of a (pure) cryptocurrency pair.}
\usage{
poloniex_download(
  from = "USDT",
  to = "ETH",
  start = "2001-01-01",
  end = "2017-01-01",
  period = 86400
)
}
\arguments{
\item{from}{is the base currency of the currency pair}

\item{to}{is the quote currency of the currency pair}

\item{start}{is the starting point of the required data sample}

\item{end}{is the ending point of the required data sample}

\item{period}{is the candlestick period in seconds (valid values are 300, 900, 1800, 7200, 14400, and 86400)}
}
\value{
a dataframe with data of HLOCV of the referred cryptocoin
}
\description{
This function downloads time series data of a (pure) cryptocurrency pair (i.e. crypto vs crypto) from the Poloniex exchange,
given a specified starting point, ending point and candlestick period in seconds
}
\examples{
eth_data <- poloniex_download(from="USDT", to="ETH", start="2013-01-01", end="2017-05-18")
}
