% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liquidity_cryptowat.R
\name{prepareBestOrders}
\alias{prepareBestOrders}
\title{Extract best bids and best asks from the orderbook.csv file}
\usage{
prepareBestOrders(market, pair, sourceDirectory)
}
\arguments{
\item{market}{is the required exchange}

\item{pair}{is the required exchange rate pair}

\item{sourceDirectory}{is the directory path where the csv order-book data files is located
(and where the best_orders.csv file will be saved)}
}
\description{
This function extract best bids and best asks from the orderbook.csv file
}
\details{
This function extracts best bids and best asks from the orderbook.csv file and save them in the best_orders.csv file
}
\examples{
\dontrun{
data_file<-system.file("extdata", package = "bitcoinFinance")
prepareOrders("coinbase","btcusd",data_file,data_file)
prepareBestOrders("coinbase","btcusd",data_file )
}
}
