% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liquidity_cryptowat.R
\name{prepareOrders}
\alias{prepareOrders}
\title{Extract the downloaded cryptocurriency order-book data from the json files and save them into a csv file}
\usage{
prepareOrders(market, pair, sourceDirectory, destDirectory)
}
\arguments{
\item{market}{is the required exchange}

\item{pair}{is the required exchange rate pair}

\item{sourceDirectory}{is the directory path where the json order-book data files are located}

\item{destDirectory}{is the directory path where the csv order-book files will be saved}
}
\description{
This function extracts the downloaded cryptocurriency order-book data from the json files and save them into a csv file
}
\details{
This function extracts the downloaded cryptocurriency order-book data from the json files and save them into a csv file
}
\examples{
\dontrun{
data_file<-system.file("extdata", package = "bitcoinFinance")
prepareOrders("coinbase","btcusd",data_file,data_file)
}
}
