% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positive_definite_realized_covariances.R
\name{rc_pd}
\alias{rc_pd}
\title{Compute the realized covariance of an xts object containing bitcoin intraday prices and regularize them if not positive definite}
\usage{
rc_pd(price_xts)
}
\arguments{
\item{price_xts}{is a xts object containing intraday bitcoin prices (2 series or more)}
}
\value{
final_rc a list containing the (potentially regularized) realized covariances
}
\description{
This function computes the realized covariance of an xts object containing bitcoin intraday prices and regularize them if not positive definite
}
\details{
This function computes the realized covariance of an xts object containing
bitcoin intraday prices and regularize them if not positive definite, as done in
Hautsch, N., L. Kyj, and R. Oomen (2012)
}
\examples{
\dontrun{
library(highfrequency)
localbtcMXN <-bitcoincharts_single_download(name = "localbtcMXN.csv.gz")
btcdeEUR <-bitcoincharts_single_download(name = "btcdeEUR.csv.gz")
krakenEUR <-bitcoincharts_single_download(name = "krakenEUR.csv.gz")
localbtcINR <-bitcoincharts_single_download(name = "localbtcINR.csv.gz")
coinbaseUSD <-bitcoincharts_single_download(name = "coinbaseUSD.csv.gz")
data5<-list(localbtcMXN =localbtcMXN, btcdeEUR=btcdeEUR, krakenEUR=krakenEUR,
             localbtcINR=localbtcINR, coinbaseUSD=coinbaseUSD)
data_clean<-aggregate_merge_bictoincharts_data(data_list=data5)
cleaned.rc <- rc_pd(data_clean$price_ts)
}

}
\references{
Hautsch, N., L. Kyj, and R. Oomen (2012). A blocking and regularization approach
to high-dimensional realized covariance estimation. Journal of Applied Econometrics, 27, 625–645
}
