% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{recursive.arch.multi}
\alias{recursive.arch.multi}
\title{Compute a sequence of  multivariate ARCH tests}
\usage{
recursive.arch.multi(resids, max.lag)
}
\arguments{
\item{resids}{is a dataframe containing the standardized residuals from a multivariate model}

\item{max.lag}{is the the max lag used to compute the multivariate ARCH test}
}
\value{
results is a list with the lag, statistic and p-values
}
\description{
This function computes a sequence of multivariate ARCH tests
}
\details{
This function computes a sequence of multivariate ARCH tests using a modified version of
the original arch.multi internal function of the vars package
}
\examples{
\dontrun{
 Sigma <- matrix(c(10,3,3,2),2,2)
 dat.norm<- MASS::mvrnorm(n = 1000, rep(0, 2), Sigma)
 recursive.arch.multi(dat.norm,20)
}
}
