% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{recursive.portmanteau.multi}
\alias{recursive.portmanteau.multi}
\title{Compute a sequence of  multivariate Portmanteau tests}
\usage{
recursive.portmanteau.multi(resids, var.p.lag, max.lag)
}
\arguments{
\item{resids}{is a dataframe containing the standardized residuals from a multivariate model}

\item{var.p.lag}{is the number of lags in the estimated multivariate VAR(p) model}

\item{max.lag}{is the the max lag used to compute the multivariate Portmanteau  test}
}
\value{
results is a list with the lag, statistic and p-values
}
\description{
This function computes a sequence of multivariate Portmanteau tests
}
\details{
This function computes  a sequence of multivariate Portmanteau tests using a modified version of
the original portmanteau.multi internal function of the vars package
}
\examples{
\dontrun{
 Sigma <- matrix(c(10,3,3,2),2,2)
 dat.norm<- MASS::mvrnorm(n = 1000, rep(0, 2), Sigma)
 var.2c <- vars::VAR(dat.norm, p = 1, type = "const")
 res.var<-resid(var.2c)
 recursive.portmanteau.multi(res.var,1,20)
}
}
