% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio_management_crypto.R
\name{W_sd_sr_out}
\alias{W_sd_sr_out}
\title{Compute the optimal power-law parameter and the portfolio weights according to the approach proposed by Kristoufek (2013) - [out-of-sample]}
\usage{
W_sd_sr_out(V, alpha, ret_data, rf = 0.005, type = "StdDev", lag_ws = 1)
}
\arguments{
\item{V}{a zoo object containing the Google Trends search terms (cryptocurrency ticker or other type)}

\item{alpha}{is a vector of possible values for the power-law parameter alpha (usually between -2 and +2)}

\item{ret_data}{a zoo object containing the return data of the selected cryptocurrencies}

\item{rf}{is the risk-free rate needed for the computation of the \link[PerformanceAnalytics]{SharpeRatio}}

\item{type}{one of "StdDev" or "VaR" or "ES" to use as the denominator for computing the \link[PerformanceAnalytics]{SharpeRatio}}

\item{lag_ws}{the lag to be applied to the portfolio weights for computing the out-of-sample analysis}
}
\value{
a list containing the following objects:
\describe{
  \item{sd_all}{a vector of all standard deviations obtained in-sample for each value of alpha}
  \item{min_sd}{value of alpha that min the standard deviation in-sample}
  \item{sr_all}{a vector of all Sharpe Ratios obtained in-sample for each value of alpha}
  \item{min_sr}{value of alpha that max the Sharpe Ratio in-sample}
}
}
\description{
This function computes the optimal power-law parameter and the portfolio weights according to the approach proposed by Kristoufek (2013)
using out-of-sample data. Default lag for portfolio weights is 1, but it can be modified by the user.
}
\details{
This function computes the optimal power-law parameter alpha that min the std. dev or max the Sharpe ratio for the given dataset,
using (pseudo) out-of-sample data. The portfolio weights (computed for a given alpha) are lagged by a quantity
defined by \code{lag_ws}.
}
\examples{
\dontrun{
library(PerformanceAnalytics)
data(crypto_google_data)
R <- crypto_google_data[,2:6]
R<- zoo::zoo( apply(R, 2, QuantTools::returns), order.by=lubridate::ymd(crypto_google_data[,1]))
R<-R[2:nrow(R),]
alpha <- seq(from = -2, to = 2, by = 0.1)
V=zoo::zoo(crypto_google_data[,7:11], order.by=lubridate::ymd(crypto_google_data[,1]) )
V<-V[2:nrow(V),]
google_port_out=W_sd_sr_out(V, alpha, ret_data=R, rf=0.003, type="StdDev", lag_ws=4)
#Plots
plot(y = google_port_out$sd_all,
     x = alpha,
     xlab = "Discrimination parameter",
     ylab = "StdDev",pch=21)
plot(y = google_port_out$sr_all,
     x = alpha,
     xlab = "Discrimination parameter",
     ylab = "SharpeRatio")
     }

}
\references{
Kristoufek, L. (2013). Can Google Trends search queries contribute to risk diversification?. Scientific reports, 3, 2713.
}
