% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zpp.R
\name{zpp_general}
\alias{zpp_general}
\title{Compute the general version of the ZPP}
\usage{
zpp_general(prices, mean.model = list(armaOrder = c(0, 0)),
  variance.model = list(garchOrder = c(1, 1)),
  distribution.model = "norm", n.days = 250, scenarios = 20000)
}
\arguments{
\item{prices}{is a numeric vector containing the prices series}

\item{mean.model}{is a list specifying the ARMA order for \code{\link{ugarchspec}}}

\item{variance.model}{is a list containing the variance model specification for \code{\link{ugarchspec}}}

\item{distribution.model}{is the conditional density to use for the innovations. See \code{\link{ugarchspec}} for details}

\item{n.days}{is the number of time-steps}

\item{scenarios}{is the total number fo trajectories}
}
\value{
zpp is the estimated ZPP
}
\description{
This function computes the general version of the ZPP by simulating a number of trajectories (total number is given
by \code{scenarios}) over \code{n.days} time-steps, given an estimated GARCH model
}
\examples{
eth_data<-poloniex_download(from="USDT",to="ETH",start="2001-01-01",end="2017-05-18",period=86400)
prices<-eth_data$weightedAverage
zpp<-zpp_general(prices)
zpp
}
