% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zpp.R
\name{zpp_norm}
\alias{zpp_norm}
\title{Compute the closed-form normal ZPP}
\usage{
zpp_norm(prices, n.days = 250)
}
\arguments{
\item{prices}{is a numeric vector containing the prices series}

\item{n.days}{is the number of time-steps}
}
\value{
zpp.norm is the estimated closed-form normal ZPP
}
\description{
This function computes the closed-form normal ZPP assuming the price series follows a random walk with drift with normal innovations
}
\examples{
eth_data<-download_coinmarketcap_daily(id=1027,currency="USD",start_date="2016-09-16",
          end_date="2017-09-15")
prices<-eth_data$close
zpp.norm<-zpp_norm(prices)
zpp.norm
}
