% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_bound.R
\name{btc.lower.bound.web}
\alias{btc.lower.bound.web}
\title{Compute the bitcoin lower bound with web-delivered inputs}
\usage{
btc.lower.bound.web(
  block.reward = 12.5,
  hashing.power.miner = 10^12,
  W.GHs = 0.25
)
}
\arguments{
\item{block.reward}{is the block reward (currently 12.5 BTC/block)}

\item{hashing.power.miner}{is the hashing power employed by a miner. Hayes (2017) sets this value to 1000 GH/s even though
the actual hashing power of a miner is likely to deviate greatly from this value.}

\item{W.GHs}{is the energy consumption efficiency of the producer’s hardware expressed in W per GH/s}
}
\value{
price.bitcoin is the bitcoin lower bound
}
\description{
This function computes the bitcoin lower bound with (some) inputs taken from the web
}
\details{
This function computes the bitcoin lower bound with some inputs taken from the web: a) the latest difficulty (expressed in units of GH/block),
b) the electricity price per kWh in dollars (given by the latest US average retail price - residential sector)
}
\examples{
 btc.lower.bound.web(W.GHs=0.2)

}
\references{
Hayes, A. S. (2017). Cryptocurrency Value Formation: An empirical study leading to a cost of production model for valuing Bitcoin. Telematics and Informatics, forthcoming.
}
