% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_coinmarketcap.R
\name{download_coinmarketcap_w}
\alias{download_coinmarketcap_w}
\title{Download the weekly data of the selected cryptocurrency from coinmarketcap.com}
\usage{
download_coinmarketcap_w(
  id = 1,
  currency = "USD",
  start_date = "2018-09-16",
  end_date = "2020-09-23"
)
}
\arguments{
\item{id}{is the coinmarketcap ID used for the selected cryptocurrency (see the coinmarketcap API for the full ID list)}

\item{currency}{is the reference currency used in the crypto pair}

\item{start_date}{is the starting date (format YYYY-MM-DD) for downloading the selected cryptocurrency weekly data}

\item{end_date}{is the ending date (format YYYY-MM-DD) for downloading the selected cryptocurrency weekly data}
}
\value{
table_news is a dataframe with weekly open high low close volume and market capitalization
}
\description{
This function downloads the weekly data of the selected cryptocurrency from coinmarketcap.com
}
\details{
This function downloads the weekly data of the selected cryptocurrencies from coinmarketcap.com
}
\examples{
dat<-download_coinmarketcap_w(id=2,currency="USD",start_date="2016-09-16",end_date="2020-09-23")
head(dat)

}
