% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zpp.R
\name{zpp_norm}
\alias{zpp_norm}
\title{Compute the closed-form normal ZPP}
\usage{
zpp_norm(prices, n.days = 250)
}
\arguments{
\item{prices}{is a numeric vector containing the prices series}

\item{n.days}{is the number of time-steps}
}
\value{
zpp.norm is the estimated closed-form normal ZPP
}
\description{
This function computes the closed-form normal ZPP assuming the price series follows a random walk with drift with normal innovations
}
\examples{
btc_data<-dat<-bitcoinity_download(currency="USD", data_type="price",
               exchange="coinbase",time_length = "2y")
prices<-btc_data$avg
zpp.norm<-zpp_norm(prices)
zpp.norm
}
